Refactored static file serving
The HTTPWriter is now available to the context. This allows static
file serving to behandled completely by the LocalFileRoute and
LocalPathRoute. Filepaths no longer need to be passed through the
response into the ServeHTTPMethod. The route now responds with a
response with Handled set to true skipping the logic in the ServeHTTP
function.

Filepath and Filename attributes were removed from the Response
structure and the File function was removed from the Context.
