+++ response.go
@@ -12,8 +12,6 @@ type Response struct {
 	Error      error
 	Meta       map[string]interface{}
 	Header     http.Header
-	Filepath   string
-	Fileroot   string
 	raw        []byte
 	Handled    bool
 }
@@ -46,11 +44,6 @@ func (r *Response) Success() bool {
 	return r.Error == nil
 }
 
-// IsFile returns true if the response should output a local file
-func (r *Response) IsFile() bool {
-	return (r.Filepath != "")
-}
-
 // IsRaw determens if the response is a raw response
 func (r *Response) IsRaw() bool {
 	return len(r.raw) > 0
@@ -59,7 +52,6 @@ func (r *Response) IsRaw() bool {
 // SetRaw sets the responses raw output
 func (r *Response) SetRaw(b []byte) {
 	r.raw = b
-	r.Filepath = ""
 	r.Data = nil
 }
