+++ route.go
@@ -2,8 +2,6 @@ package celerity
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
 )
 
 // Route is an interface that can be implemented by any objects wishing to
@@ -42,75 +40,4 @@ func (r *BasicRoute) RoutePath() RoutePath {
 
 func (r *BasicRoute) String() string {
 	return fmt.Sprint(r.Method, "\t", r.Path)
-}
-
-// LocalFileRoute will handle serving a single file from the local file system
-// to a given path.
-type LocalFileRoute struct {
-	Path      RoutePath
-	LocalPath string
-}
-
-// Match checks if the incoming path matches the route path and if the local
-// file exists.
-func (l *LocalFileRoute) Match(method string, path string) (bool, string) {
-	ok, xtra := l.Path.Match(path)
-	fs := FSAdapter.RootPath(filepath.Dir(l.LocalPath))
-	if !ok || method != GET || xtra != "" {
-		return false, path
-	}
-	fname := "/" + filepath.Base(l.LocalPath)
-	if _, err := fs.Stat(fname); os.IsNotExist(err) {
-		return false, path
-	}
-	return true, xtra
-}
-
-// Handle sets the response to return a local file.
-func (l *LocalFileRoute) Handle(c Context) Response {
-	fname := "/" + filepath.Base(l.LocalPath)
-	fpath := filepath.Dir(l.LocalPath)
-	return c.File(fpath, fname)
-}
-
-// RoutePath returns the route path for the route.
-func (l *LocalFileRoute) RoutePath() RoutePath {
-	return l.Path
-}
-
-// LocalPathRoute handles serving any file under a path. If the requested file
-// exists it will be served if not the router will continue processing
-// routes.
-type LocalPathRoute struct {
-	Path      RoutePath
-	LocalPath string
-}
-
-// Match checks if a file exists under the local path
-func (l *LocalPathRoute) Match(method string, path string) (bool, string) {
-	fs := FSAdapter.RootPath(l.LocalPath)
-	if len(path) < len(l.Path) {
-		return false, path
-	}
-	fname := "/" + path[len(l.Path):]
-	stat, err := fs.Stat(fname)
-	if err != nil {
-		return false, path
-	}
-	if stat.Mode().IsDir() {
-		return false, path
-	}
-	return true, ""
-}
-
-// Handle sets the resposne up to serve the local file.
-func (l *LocalPathRoute) Handle(c Context) Response {
-	fname := c.ScopedPath[len(l.Path):]
-	fpath := l.LocalPath
-	return c.File(fpath, fname)
-}
-
-// RoutePath returns the routepath for the route
-func (l *LocalPathRoute) RoutePath() RoutePath {
-	return l.Path
 }
